refactor dashboard components
